Merchify_Backend/requirements.txt
djangorestframework
django-cors-headers
pillow
jsonschema
